% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GSTobj.R
\name{as.GST}
\alias{as.GST}
\title{as Group Sequential Trial}
\usage{
as.GST(GSD, GSDo)
}
\arguments{
\item{GSD}{object of the \code{class} \code{GSTobj}; group sequential design}

\item{GSDo}{group sequential design outcome;  a list with the variables \code{T} and \code{z}; list(T = stage where trial stops, z = z-statistic at stage where trial stops)}
}
\value{
Returns a list containing the \code{GSD} and \code{GSDo} with \code{class}=\code{GSTobj}
}
\description{
Function \code{as.GST} builds a group sequential trial object
}
\examples{
GSD <- plan.GST(K=4,SF=1,phi=0,alpha=0.025,delta=6,pow=0.8,compute.alab=TRUE,compute.als=TRUE)
GSDo <- list(T=2, z=3.1)
GST <- as.GST(GSD=GSD,GSDo=GSDo)
GST
}
\author{
Niklas Hack \email{niklas.hack@meduniwien.ac.at} and Werner Brannath \email{werner.brannath@meduniwien.ac.at}
}
\seealso{
\code{\link{GSTobj}}
}
\keyword{methods}

